refactor: Move mobile state to MobileNav
